Changing the final responder to return a 444 No Response, instead of a 404 not found. That can then be picked up and replaced with a 404 using a default interceptResponse directive. Which means if two middlewares (like input-nextjs and serve-static) both want to return a 404, they don’t clobber each other.
