+++ packages/core/src/constants.ts
@@ -1,6 +1,6 @@
 export const DEFAULT_CONFIG_FILENAME = 'fab.config.json5'
-
 export const DEFAULT_MIME_TYPE = 'text/html; charset=utf-8'
+export const NO_RESPONSE_STATUS_CODE = 444
 
 export const IMMUTABLE_HEADERS = { 'Cache-Control': 'immutable' }
 export const NON_IMMUTABLE_HEADERS = { 'Cache-Control': 'no-cache' }
